Api change /achievements/grouped -> /achievements?layout=grouped
